feat: 아이템 리스트 렌더링 성능 최적화 및 캐시 필터 복구
- React.memo를 활용한 ItemCard 컴포넌트 분리 및 최적화
- ItemList 컴포넌트 메모이제이션 적용
- 이미지 lazy loading 구현으로 초기 로딩 성능 개선
- useCallback으로 이벤트 핸들러 최적화
- 캐시 아이템 필터 토글 버튼 복구 (장비 카테고리 전용)
- useEffect 의존성 배열 오류 수정
- UI 자연스럽게 통합 (별도 섹션 제거)
